LAPA/requirements.txt
git+https://github.com/lhao499/tux.git
flax==0.7.0
optax==0.1.7
chex==0.1.82
einops
transformers==4.29.2; python_version < "3.12"
transformers>=4.41,<4.48; python_version >= "3.12"
datasets==2.13.0
tqdm
ml_collections
wandb
gcsfs
requests
typing-extensions
sentencepiece
Pillow
ipdb
imageio[ffmpeg]
decord
tiktoken
jax==0.4.23
jaxlib==0.4.23
tensorflow==2.15.1; python_version < "3.12"
tensorflow>=2.16,<2.18; python_version >= "3.12"
scipy==1.12.0
albumentations
uvicorn
fastapi